requirements.txt
google-adk[a2a]>=1.0.0
a2a-sdk>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
httpx-sse>=0.4.0
python-dotenv>=1.0.0
